Moved the group-by-status table and helper functions into a separate component.
